backend/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
python-dotenv==0.13.0
pytz==2019.3
sqlparse==0.3.1
